The performance of most typical classifier learning methods, which assume a generally  balanced class distribution and equal misclassification costs, has been severely hampered when  it comes to classifying data with an uneven class distribution. Data imbalance is an unavoidable  element of real-time data that must be addressed. Experiments show that data imbalance and  data size have an impact on predictors. A performance analysis of several standard algorithms  was carried out in order to get accurate result using different methods and techniques.
If you've worked in machine learning or data science for any length of time, you've almost
certainly come across imbalanced class distribution. This is a situation in which the number of
observations in one class is much lower than the number of observations in the other classes.
This issue is particularly prevalent in situations where anomaly detection is critical, such as
electricity theft, fraudulent financial transactions, and the diagnosis of rare diseases, among
others. In this case, the predictive model built with traditional machine learning methods may
be biased and erroneous.
Because Machine Learning Algorithms are typically designed to improve accuracy by reducing
error, this happens. As a result, they do not account for class distribution, proportion, or
balance.
Using various sampling strategies, this tutorial covers various approaches for tackling such
class imbalance problems. We also consider the advantages and disadvantages of each
technique. Finally, I show how to generate a balanced class distribution using an ensemble
learning technique that was created specifically for this purpose.